requirements.txt
asgiref==3.8.1
black==24.4.1
click==8.1.7
Django==5.0
django-environ==0.11.2
djangorestframework==3.15.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
psycopg2==2.9.9
sqlparse==0.5.0
